refactor(misconf): move node transformation logic to cf package
