+++ pkg/iac/ast/yaml.go
@@ -4,7 +4,6 @@ import (
 	"encoding/base64"
 	"fmt"
 	"strconv"
-	"strings"
 	"time"
 
 	"gopkg.in/yaml.v3"
@@ -13,31 +12,9 @@ import (
 	"github.com/aquasecurity/trivy/pkg/x/json"
 )
 
-const mapTag = "!!map"
-
-func isCustomTag(tag string) bool {
-	return strings.HasPrefix(tag, "!") && !strings.HasPrefix(tag, "!!")
-}
-
 func (n *Node) UnmarshalYAML(node *yaml.Node) error {
 	n.StartLine = node.Line
 	n.EndLine = calculateEndLine(node)
-
-	if isCustomTag(node.Tag) {
-		var newContent []*yaml.Node
-
-		newContent = append(newContent, &yaml.Node{
-			Tag:   "!!str",
-			Value: getIntrinsicTag(node.Tag),
-			Kind:  yaml.ScalarNode,
-		})
-
-		newContent = createNode(node, newContent)
-
-		node.Tag = mapTag
-		node.Kind = yaml.MappingNode
-		node.Content = newContent
-	}
 
 	switch node.Tag {
 	case "!!string", "!!str":
@@ -78,7 +55,7 @@ func (n *Node) UnmarshalYAML(node *yaml.Node) error {
 		}
 		n.Value = val
 		n.Kind = BinaryNode
-	case mapTag:
+	case "!!map":
 		entries, err := handleMapTag(node)
 		if err != nil {
 			return err
@@ -101,45 +78,6 @@ func (n *Node) UnmarshalYAML(node *yaml.Node) error {
 		)
 	}
 	return nil
-}
-
-// TODO: use custom unmarshaller
-func createNode(node *yaml.Node, newContent []*yaml.Node) []*yaml.Node {
-	if node.Content == nil {
-		newContent = append(newContent, &yaml.Node{
-			Tag:   "!!str",
-			Value: node.Value,
-			Kind:  yaml.ScalarNode,
-		})
-	} else {
-
-		newNode := &yaml.Node{
-			Content: node.Content,
-			Kind:    node.Kind,
-		}
-
-		switch node.Kind {
-		case yaml.SequenceNode:
-			newNode.Tag = "!!seq"
-		case yaml.MappingNode:
-			newNode.Tag = mapTag
-		case yaml.ScalarNode:
-		default:
-			newNode.Tag = node.Tag
-		}
-		newContent = append(newContent, newNode)
-	}
-	return newContent
-}
-
-func getIntrinsicTag(tag string) string {
-	tag = strings.TrimPrefix(tag, "!")
-	switch tag {
-	case "Ref", "Contains":
-		return tag
-	default:
-		return fmt.Sprintf("Fn::%s", tag)
-	}
 }
 
 func handleMapTag(node *yaml.Node) (map[string]*Node, error) {
